+++ indyscan-webapp/pages/home.js
@@ -1,27 +1,25 @@
 import React, { Component } from 'react'
 import '../scss/style.scss'
-import TxsChart from '../components/TxChart/TxChart'
 import { getTransactions, getTxTimeseries, getNetwork } from 'indyscan-api'
 import { getBaseUrl } from '../routing'
 import { Grid, GridColumn, GridRow } from 'semantic-ui-react'
 import PageHeader from '../components/PageHeader/PageHeader'
 import TxPreviewList from '../components/TxPreviewList/TxPreviewList'
 import Footer from '../components/Footer/Footer'
-import IntervalDropdown from '../components/IntervalDropdown/IntervalDropdown'
+import fetch from 'isomorphic-fetch'
 
-async function fetchTimeseriesData(baseUrl, network, utimeSince, utimeUntil) {
-  const timeseriesDomain = (await getTxTimeseries(baseUrl, network, 'domain', 'auto', utimeSince, utimeUntil)).map(m => [m[0] * 1000, m[1]])
-  const timeseriesPool = (await getTxTimeseries(baseUrl, network, 'pool', 'auto', utimeSince, utimeUntil)).map(m => [m[0] * 1000, m[1]])
-  const timeseriesConfig = (await getTxTimeseries(baseUrl, network, 'config', 'auto', utimeSince, utimeUntil)).map(m => [m[0] * 1000, m[1]])
-  return {timeseriesDomain, timeseriesPool, timeseriesConfig}
-}
+function secondsToDhms(seconds) {
+  seconds = Number(seconds);
+  var d = Math.floor(seconds / (3600*24));
+  var h = Math.floor(seconds % (3600*24) / 3600);
+  var m = Math.floor(seconds % 3600 / 60);
+  var s = Math.floor(seconds % 60);
 
-const secsInDay = 3600 * 24
-const secsInWeek = secsInDay * 7
-
-function utimeWeekAgo() {
-  const utimeSecNow = Math.floor(new Date() / 1000)
-  return (utimeSecNow - secsInWeek)
+  var dDisplay = d > 0 ? d + (d == 1 ? " day, " : " days, ") : "";
+  var hDisplay = h > 0 ? h + (h == 1 ? " hour, " : " hours, ") : "";
+  var mDisplay = m > 0 ? m + (m == 1 ? " minute, " : " minutes, ") : "";
+  var sDisplay = s > 0 ? s + (s == 1 ? " second" : " seconds") : "";
+  return dDisplay + hDisplay + mDisplay + sDisplay;
 }
 
 class HomePage extends Component {
@@ -30,85 +28,31 @@ class HomePage extends Component {
     const baseUrl = getBaseUrl(req)
     const { network } = query
     const networkDetails = await getNetwork(baseUrl, network)
-    console.info(`[getInitialProps:HomePage]... network = ${network}`)
-    console.info(`[getInitialProps:HomePage]... network details = ${JSON.stringify(networkDetails)}`)
     const domainTxs = await getTransactions(baseUrl, network, 'domain', 0, 13)
-    const utimeSecNow = Math.floor(new Date() / 1000)
-    const {timeseriesDomain, timeseriesPool, timeseriesConfig} = await fetchTimeseriesData(baseUrl, network, utimeWeekAgo(), utimeSecNow)
+    const poolTxs = await getTransactions(baseUrl, network, 'pool', 0, 13)
+    const configTxs = await getTransactions(baseUrl, network, 'config', 0, 13)
+    const versionRes = await fetch(`${baseUrl}/version`)
+    const version = (await versionRes.json()).version
     return {
       networkDetails,
       network,
-      txs: domainTxs.txs,
-      timeseriesDomain,
-      timeseriesPool,
-      timeseriesConfig,
+      domainTxs: domainTxs.txs,
+      poolTxs: poolTxs.txs,
+      configTxs: configTxs.txs,
       baseUrl,
-      timerangeSec: secsInWeek
+      version
     }
   }
 
-  getTimerangeSec() {
-    const fromState = (this.state && this.state.timerangeSec)
-    console.log(`Returning timerange from state? ${fromState}`)
-    const value = (fromState)
-      ? this.state.timerangeSec
-      : this.props.timerangeSec
-    console.log(`Determined timerange: ${value}`)
-    return value
-  }
-
-  getTimeseriesRenderData() {
-    const renderFromState = ((!!this.state) && (!!this.state.timeseriesDomain) && (!!this.state.timeseriesPool) && (!!this.state.timeseriesConfig))
-    console.log(`renderFromState data from state ${renderFromState}`)
-    if (renderFromState) {
-      const {timeseriesDomain, timeseriesPool, timeseriesConfig} = this.state
-      return {timeseriesDomain, timeseriesPool, timeseriesConfig}
-    } else {
-      const {timeseriesDomain, timeseriesPool, timeseriesConfig} = this.props
-      return {timeseriesDomain, timeseriesPool, timeseriesConfig}
-    }
-  }
-
-  async reloadGraphDataToState (timerangeSec) {
-    const utimeSecNow = Math.floor(new Date() / 1000)
-    const sinceUtimeSec = utimeSecNow - timerangeSec
-    console.log(`Will fetch series data for length of ${(utimeSecNow-sinceUtimeSec) / (3600 * 24)} days`)
-    const {baseUrl, network} = this.props
-    const {timeseriesDomain, timeseriesPool, timeseriesConfig} = await fetchTimeseriesData(baseUrl, network, sinceUtimeSec, utimeSecNow)
-    this.setState({ timeseriesDomain })
-    this.setState({ timeseriesPool })
-    this.setState({ timeseriesConfig })
-  }
-
-  componentWillReceiveProps (nextProps, nextContext) {
-    // I suppose not the most beautiful solution but good enough for now.
-    // A page might have existing state as user is changing the timerange displayed on graph
-    // However, if then user click on different network, that will invoke getInitialProps and componentDidUpdate
-    // Now even if I fetch new data in getInitialProps, if the previous state (the graph data) is not modified,
-    // the new data will be in props, but the state will stay the old. Because the page gives precedence to display
-    // state data instead of props, it will keep displaying old data from previous network, until the user causes
-    // state
-    this.setState({timeseriesDomain:null})
-    this.setState({timeseriesPool:null})
-    this.setState({timeseriesConfig:null})
-    this.setState({timerangeSec:secsInWeek})
-  }
-
-
-  setGraphTimerange (timerangeSec) {
-    this.setState({timerangeSec})
-    this.reloadGraphDataToState(timerangeSec)
+  calculateTimeSinceLastTransaction = function calculateTimeSinceLastTransaction(txs) {
+    const timestamps = txs.map(tx => (tx && tx.txnMetadata) ? tx.txnMetadata.txnTime : undefined).filter(t => !!t)
+    const utimeMaxTx = Math.max(...timestamps)
+    const utimeNow = Math.floor(new Date() / 1000)
+    return secondsToDhms(utimeNow-utimeMaxTx)
   }
 
   render () {
-    if (!(this.state) || !(this.state.timerangeSec)) {
-      console.log(`Rendering and PROPS timerange is ${this.props.timerangeSec}`)
-    } else {
-      console.log(`Rendering when timerange STATE value is ${this.state.timerangeSec}`)
-    }
     const { network, networkDetails, baseUrl } = this.props
-    const {timeseriesDomain, timeseriesPool, timeseriesConfig} = this.getTimeseriesRenderData()
-    const timerangeSec = this.getTimerangeSec()
     console.log(JSON.stringify(networkDetails))
     return (
       <Grid>
@@ -126,31 +70,36 @@ class HomePage extends Component {
           </p>
         </GridRow>
         <GridRow>
-          <GridColumn width={10} floated='left'>
-            <Grid>
-              <GridRow>
-                <IntervalDropdown value={timerangeSec}
-                  changeGraphTimerange={this.setGraphTimerange.bind(this)}></IntervalDropdown>
-              </GridRow>
-              <GridRow>
-                <GridColumn>
-                        <TxsChart label="Domain tx count" color="darkcyan" type="domain"
-                                  timeseries={timeseriesDomain}/>
-                        <TxsChart label="Pool tx count" color="mediumaquamarine" type="pool"
-                                  timeseries={timeseriesPool}/>
-                        <TxsChart label="Config tx count" color="dodgerblue" type="config"
-                                  timeseries={timeseriesConfig}/>
-                </GridColumn>
-              </GridRow>
-            </Grid>
-          </GridColumn>
-          <GridColumn width={6} floated='right' style={{ paddingLeft: '7em' }}>
-            <GridRow centered>
-              <h2>Last domain transactions</h2>
+          <GridColumn width={5} align='left'>
+            <GridRow align='left'>
+              <h2>Domain txs</h2>
+              <h4>Last tx {this.calculateTimeSinceLastTransaction(this.props.domainTxs)} ago</h4>
             </GridRow>
             <GridRow centered style={{ marginTop: '2em' }}>
               <Grid.Column>
-                <TxPreviewList txs={this.props.txs} network={network}/>
+                <TxPreviewList txs={this.props.domainTxs} network={network} subledger='domain'/>
+              </Grid.Column>
+            </GridRow>
+          </GridColumn>
+          <GridColumn width={6} align='center'>
+            <GridRow align='left'>
+              <h2>Pool txs</h2>
+              <h4>Last tx {this.calculateTimeSinceLastTransaction(this.props.poolTxs)} ago</h4>
+            </GridRow>
+            <GridRow centered style={{ marginTop: '2em' }}>
+              <Grid.Column>
+                <TxPreviewList txs={this.props.poolTxs} network={network} subledger='pool'/>
+              </Grid.Column>
+            </GridRow>
+          </GridColumn>
+          <GridColumn width={5} align='right'>
+            <GridRow align='left'>
+              <h2>Config txs</h2>
+              <h4>Last tx {this.calculateTimeSinceLastTransaction(this.props.configTxs)} ago</h4>
+            </GridRow>
+            <GridRow centered style={{ marginTop: '2em' }}>
+              <Grid.Column>
+                <TxPreviewList txs={this.props.configTxs} network={network} subledger='config'/>
               </Grid.Column>
             </GridRow>
           </GridColumn>
@@ -158,7 +107,7 @@ class HomePage extends Component {
         </GridRow>
         <GridRow>
           <GridColumn>
-            <Footer/>
+            <Footer displayVersion={this.props.version}/>
           </GridColumn>
         </GridRow>
       </Grid>