+++ indyscan-webapp/txtools/index.js
@@ -4,8 +4,9 @@ export function extractTxInformation (tx) {
   const { rootHash } = tx
   const { type } = tx.txn
   const { txnId, txnTime, seqNo } = tx.txnMetadata
+  const from = tx.txn.metadata ? tx.txn.metadata.from : 'not-available'
   const timestamp = txnTime
     ? moment(txnTime * 1000).format('MMMM Do YYYY, h:mm:ss a')
     : 'unknown'
-  return { txnId, txnTime, seqNo, timestamp, type, rootHash }
+  return { txnId, txnTime, seqNo, timestamp, type, rootHash, from }
 }