+++ changelog.md
@@ -9,6 +9,11 @@ transaction consumer.
 - Add support for new transaction types `REVOC_REG_DEF`, `REVOC_REG_ENTRY`, `AUTH_RULE`, 
 `AUTH_RULES`, `TXN_AUTHOR_AGREEMENT`, `TXN_AUTHOR_AGREEMENT_AML`
 - Use refactored storage module
+- Display timeseries graphs on separate page
+- Display latest transactions on homepage
+- Adjust layout of network / menu links
+- Display pretty error if tx page for particular transaction can't be displayed
+- Display current version of webapp at footer of homepage
 
 ## 1.x.x
 - First official release